backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.28.1